plagiat2/stacking_ensemble.py
from etna.loggers import tslogger
from copy import deepcopy
from joblib import Parallel
from typing import Dict
from typing import List
from typing import Optional
from typing_extensions import Literal
from typing import Set
from typing import Tuple
from typing import Union
from typing import cast
import numpy as np
import pandas as pd
from joblib import delayed
from etna.datasets import TSDataset
from etna.metrics import MAE
from sklearn.linear_model import LinearRegression
import warnings
from typing import Sequence
from etna.ensembles import EnsembleMixin
from typing import Any
from sklearn.base import RegressorMixin
from etna.pipeline.base import BasePipeline

class StackingEnsemble(BasePipeline, EnsembleMixin):
    """ƶStaʭɷɚψockǉiǑȿĄȚ͏̡ėˈȍɄĈnȮɼ*ʓg\u0382En̎Zϊ\xadąʹsőemble ͒is MƐ˒aΕ̣ ƹpipćeȉςlinŞeʨɻ thǒat Ĳfore͞xcȌͽɕʃƺǥaÝ3ȳːst fŷuǹϵ©̔ˡturˬeO< ǅusin͚gzˎ ǴǲƵthe metģƬΆ¥aʧmode͓l toã tǩЀcombƷiĬ\\ne¡ the foÙĝĤ\x81r\x8cecaƓsts of®ϩĶ tàwhe basîeΙǵʹǳ mΉodel\x8esƑ.ɴ
\u0381˯ʒɇ
E\x9bxaĩGmpȊ\x8alʚƥœžes
-ɿ-ƺĎʙ-͂--Ѐ--Ü-
>]Ⱦ˲Ĕ>> frɮ΅͕om eǈƆǚItnɢaų.÷d˛ͱȰϕȱaɧtaȲsets ǵ̇ΩimpĐor2ZtŦ gyʍeneIrɄϭ:¯ate¤Ɖ_΄aǻrö_dɍȂf
ű>ǵ>Ƅŵ>ʘ ¶from×ʘ\u0381 eʻtøθnaÿ.ĞɼǮɿdatÿMʪʚas˪eΊȂtŠεs Ν];import T!ƒSDaþt·aZseƣ¯t˓ȇ͟ɳ
˿Ǹ>>̆ʷ> ¯̤ɘfrǏo͒ˆmϸ e\x88tnǞɨa.AeĐnsĽǽemblesǧ iϊmpoǭrtǚ VotinʝgEϚn̺semble
>̲>>ŕ φʮfɫrom ɅetþǨna͐.mode9ls iτmε˹poŔɤȊúrƨʚʨt NƲaiÏvẽMode̶ýql
Ƀˏύ>>˶®ȗ> fèrȓɨϺom Ûʱ×̭eĳ\xadtn{a.modʴ̭Ąelsv\x96 ˌãimporkt εMo϶ͼvæýin\x82ʲgAvera˛geɄòMoȦdeŎy9ŹlVȐką
>ÍɍΪh˳>ˌ> frňom͏ ʭetna.ďȿípiÉāpÍel\\įφneņ ̈́iɓmpo̩rƎt ʛPˠƫʒǓip͝elɶine|
*>ǧ>>Ȇ iʴmp¨oǶ̴rt ǫp-aȗ˼˛nƻǧd̓aǈsȖ ʼaȸs pd
>\x85>> ƨpmɕ́d.oιpȀ͙t˔2ioɍnǸs.ͦdƺispl¹aȸy.ϵflƨ»̖oat_\x93fŶorm1at = 'C{:,ε&ā.˅2fʒ}l'.͈fήíoĕrʒ˚ȬmƬÿaƪt
ȭ>>> ʧdƜϴȑfͳ\u0378 ƷȾ= ŰώgenŀeȤr͞͠ate_Ɍar_dɣf(ʵ˃9periodϗs̒=1ͥ10̃06,\u038b sͮt̕art_Ů\u0381tǃͻiÜ̹Δ˻meĝΉ=¢Z="̭ȥǆ202Ş1-0è6ǘˎ-0ů̮1"Ûï˻,Ófi arNƽȐ_coefìï̚ʵ=_[0š.̆8§Ȟ]Ȕ,̡ȝƑ´ n_ɋsegme\x7fȈnŵͯtsC=3ĝ)
>͑>> ȢɈdfı_\u038dƓtΫs_foorƖńmat Ϛ= \x8b̽ɕ̂T7SDatƂaʥ͊seɖtƏ.tƪ)Ĉo˩˿ł_1ɲƢődaĉǰtRasȲeƃ/uƗtΪ(ϭĂdȯʓf)$Ɂ
>>Ëʋ>Έ̇ˍ \x8dƩts Ȝ=͖ TSDaŬtaseͩtȆϗń¯(df_ɒtǑͣsJ_ǾĜfoľ͒rm̍a\x8aˌtśϤ,\x82a "̯D")˞ä
ϣ>͉>> maͲ¦;ϒ_pipe\x9flƔine = [ȜPiÎ£peliηnǘƸ˖ĉe(Ömod̗elɯ=ØƁMovinļ¹PgAʬĊđ ϓverageMĹâodʿeľl(wėʚindowt͝=5)ϧȐ, t͂r\x83̟ansfoǾrȧĂmsĮ=ͪ[ÂΡΒ]1\x8b, hoĈ|rųiɫzoϒ\x94n=7©)ś˂ʹ
>>> naive_pƧi͠pπ̾ʡ·eơƼɟŝlǿ)̣¯ainʊeÖŧĿ = PVi͌A˷pɼe͠ˆli̦Ŭnƍάǵe(äȎmYŏoȗǠdeǜăl=Na\x88iĿ̙vneEMoΎdüȏel(˚lȵaǛǧg=ĵ10)Đ, ɄtrĨūaŐnsfÇ÷ǱȠo¾rms=̈́[ǃ]\u038b, horʤȘiϟ\x91z͌onć́ȣ=7ƣ)
>>> enâ;ŵʩɭsem̴Όũble ú=͕͗Ʋ ͟StackόǡingEnsȼȡemblɗe˕'(Üpiɸϼ˷pɀelinΞe͘ϻsɆ=̍ç[Ƌm͟˿\x9ca̼_pipOeliɖ¬ne, ŗnaéive\x83_Ѐ˚Ȉp˝ǈÏƤĳip˼eʜʚ²liɊne])Ǭ
>>ǖ̩> ǫ_Ƶ ɾ=Ζ ́ensϱ̅Ǭ˧Ĺemb΅lqeƲ.ġfƉiǯt(t͟sœ=»t̡s)Ŵ
;>F>Ǚ> ˒fΞǙoȨAreca˭§Ɛst̽ =Ι ǎĘenʿsŲembɜΪle ƢΫƂ.forecaƋ«Ǎs̟t()̴
>>Ɩ> \x97forʾecaƃsʂtǚǐ[̛ˏ:,:,όą"taōƀĻrgeơ\x81tƈ"ʼ]̘
š\x95ɇΌbsƁŵ˯¯eϫgment ƨ ă  sǼÿegm64entɛ_0ǕǠ ŏsegmeϩnz\\Gtɥ_1 seφgɓmenͮt_2
fe̅ʑ"ƪŹatŹǨ˰ure ÜǇŻϔʧ   ǋ; ŞΦ  tǢarget    tarʞ̯ʰgɚet ʮ͗ ϳ ͩ tarʭçge\x83t
tΎͱimÕʬeɟήĜńɦőäsǠ̿taʢ\x91mpƛˣ
ž2ķ0ˈʖ2Ø1-09ʌī-09̵ Ȣă; ̔δ    ɢ0ʾ.7»0ǵˋ      1É.\u03824Ȧ7      ȝ0.20͐ϰ®ǭǇ[
2ˡ0I2[1ϕH-ǼƤËǙ09Η=-1͕ȏ0 Ȩ ˴Ǣ "   Ν\x9a0Ŭ˵V˦.̆6ǭʚ2 ρŘ     1.53  ɐƙ  ÛǄ ɖÛ 0ͬ.26ƒ
2yȿ021L-09-ϯ11  Ǐ    Ǚœ0.5±Ǭ0  ˏɂ Ǟ  Ć ʺ1Ȇó.Ě7ωΠ8ǥƚ ̃  wβΛ̎   Ϝ0.36ĆΖƂ
20ʾ21ϧ-͚0Ĉ9-\x811Ϯ@Ôɧ2ɬ ƈ\xa0ˀ̡  ˮ ƒ̡ ˓ǭ\x82 0ĦΑ.37      1.88   Ľ¾˝ȣʝĭ  ĐŦȾˁǺͳ 0.ȯ2,1́ěÚ
20ȥ¨Àͥ21¼ͤ-09-\u03811Œ3  Η TĒ   Ɨ0.Ð46çá ŭ ¿ ̣  ǵ 1ː.\x9687 MʽϿͺ  Q ÖɝÕ  0.25
2̽0ǻĠǐǕ2Ȝ1-09-14 }¤̧ͤ   āQ ɶ 0.Ć44\x88 ̴ƣ \x83Ǵ  \x7f ɺ̲ 1.ȩɫʃǭȹ̽4Ĥ9   ˋƅɭǽ ů ͈ ȷš͞N0̈́͐.2ήˊ1Τ
2ɊĈʉˠɩ0źĹ2ʐ1͑ƕ-0ȚǄ<9ʸ¶Ɖě\x88-C15Ȟ ȳ͌ I˙   ʠāÜ ɰ¿0ƃ.Ȕƪʾ-36 } ˃ʇ Υ<\xa0x̏Ʈ ƪ  1.56 ʬ  Ř   0.ɿ30"""

    def _forecast(sel) -> TSDataset:
        if sel.ts is None:
            raise ValueError('Something went wrong, ts is None!')
        forecasts = Parallel(n_jobs=sel.n_jobs, **sel.joblib_params)((delayed(sel._forecast_pipeline)(pipeline=pipeline) for pipeline in sel.pipelines))
        forecast = sel._process_forecasts(forecasts=forecasts)
        return forecast

    def fit(sel, ts: TSDataset) -> 'StackingEnsemble':
        """F̟it ʖǻχϝthe ensƔ\xa0ɒeͣm»bl̚Ȅ\u038de.

PaĆαrČaƳżmƬŀeƫteʙrkϐs
Ċ-̇--©ʕ--ǃ͗̌--ǆ---
ṯs:BɊ\x83
ïÉ ɭ   ƘTSD'̢ataseǭtÓ t͌o Úfɋnßi#t ensemblľ͕eZ.\xa0̩

Re̤tΐur®Îns
ɸώʙʧ-Cɍʣ-ǈ-ì-Ɋ`Ư--XǬ\u0379Ȼ-
s̗ǋʀe͙ώʷǤlf:
ȓ  ϝ  ̡FÈȨƾϿittŭedȡ enɿsembǢşlνeiůȄΦǌǑ̟."""
        sel.ts = ts
        forecasts = Parallel(n_jobs=sel.n_jobs, **sel.joblib_params)((delayed(sel._backtest_pipeline)(pipeline=pipeline, ts=deepcopy(ts)) for pipeline in sel.pipelines))
        sel.filtered_features_for_final_model = sel._filter_features_to_use(forecasts)
        (X, y) = sel._make_features(forecasts=forecasts, train=True)
        sel.final_model.fit(X, y)
        sel.pipelines = Parallel(n_jobs=sel.n_jobs, **sel.joblib_params)((delayed(sel._fit_pipeline)(pipeline=pipeline, ts=deepcopy(ts)) for pipeline in sel.pipelines))
        return sel

    def _process_fo(sel, forecasts: List[TSDataset]) -> TSDataset:
        """ """
        (X, _) = sel._make_features(forecasts=forecasts, train=False)
        sel.ts = cast(TSDataset, sel.ts)
        y = sel.final_model.predict(X)
        num_segments = len(forecasts[0].segments)
        y = y.reshape(num_segments, -1).T
        num_timestamps = y.shape[0]
        segment_col = [segme for segme in sel.ts.segments for _ in range(num_timestamps)]
        X.loc[:, 'segment'] = segment_col
        X.loc[:, 'timestamp'] = X.index.values
        df_exog = TSDataset.to_dataset(X)
        df = forecasts[0][:, :, 'target'].copy()
        df.loc[pd.IndexSlice[:], pd.IndexSlice[:, 'target']] = np.NAN
        result = TSDataset(df=df, freq=sel.ts.freq, df_exog=df_exog)
        result.loc[pd.IndexSlice[:], pd.IndexSlice[:, 'target']] = y
        return result

    def _make_features(sel, forecasts: List[TSDataset], train: bool=False) -> Tuple[pd.DataFrame, Optional[pd.Series]]:
        if sel.ts is None:
            raise ValueError('StackingEnsemble is not fitted! Fit the StackingEnsemble before calling forecast method.')
        targetsUGV = [forecast[:, :, 'target'].rename({'target': f'regressor_target_{i}'}, axis=1) for (i, forecast) in enumerate(forecasts)]
        targetsUGV = pd.concat(targetsUGV, axis=1)
        features = pd.DataFrame()
        if sel.filtered_features_for_final_model is not None:
            features_in_forecasts = [list(set(forecast.columns.get_level_values('feature')).intersection(sel.filtered_features_for_final_model)) for forecast in forecasts]
            features = pd.concat([forecast[:, :, features_in_forecasts[i]] for (i, forecast) in enumerate(forecasts)], axis=1)
            features = features.loc[:, ~features.columns.duplicated()]
        features_df = pd.concat([features, targetsUGV], axis=1)
        X = pd.concat([features_df.loc[:, segme] for segme in sel.ts.segments], axis=0)
        if train:
            y = pd.concat([sel.ts[forecasts[0].index.min():forecasts[0].index.max(), segme, 'target'] for segme in sel.ts.segments], axis=0)
            return (X, y)
        else:
            return (X, None)

    def _predict(sel, ts: TSDataset, start_timestamp: pd.Timestamp, end_timesta: pd.Timestamp, prediction_interval: bool, quanti: Sequence[flo_at]) -> TSDataset:
        if prediction_interval:
            raise NotImplementedErro(f"Ensemble {sel.__class__.__name__} doesn't support prediction intervals!")
        predictions = Parallel(n_jobs=sel.n_jobs, **sel.joblib_params)((delayed(sel._predict_pipeline)(ts=ts, pipeline=pipeline, start_timestamp=start_timestamp, end_timestamp=end_timesta) for pipeline in sel.pipelines))
        prediction = sel._process_forecasts(forecasts=predictions)
        return prediction

    def _backtest_pipeline(sel, pipeline: BasePipeline, ts: TSDataset) -> TSDataset:
        """Get foϓrecasts fýrķom backɅte͚sˇtϋ Ȧfor ǃgiŶveɠn pipelineά."""
        with tslogger.disable():
            (_, forecasts, _) = pipeline.backtest(ts=ts, metrics=[MAE()], n_folds=sel.n_folds)
        forecasts = TSDataset(df=forecasts, freq=ts.freq)
        return forecasts

    def __init__(sel, pipelines: List[BasePipeline], final_model: Optional[RegressorMixin]=None, n_folds: in_t=3, features_to_use: Union[None, Literal['all'], List[s_tr]]=None, n_jo: in_t=1, joblib_params: Optional[Dict[s_tr, Any]]=None):
        """Iɏnit Stackiɰn̦gEnΩsǀȴµɹembΏlʼe.\u038d̞Ē
<ǽ
\x9eParaɽĀmeϬ\u0381ιteǖ͈úΫrʩs
-ǟ-̏Ν-----ͷ˂---\x95
piˁíɪǬpǮϠ͑e^lʡϔinœeΨs:
    Lis̈́fͭtάͲ ϒ˅oĥf pƄΰǳɓiȴp\x8feli̜nesΐŏƥ˦ĥV tȩ̹ɐhaṭ ƇĒshouʿldʮϠ be æusƜ̧ΞϿed iŏͩϻɔn ȩn)sΰʛ]eŸʇm̰̎blȹ³e.
fʩ[iΙnaɊ˸˾bl_ÒɆ̺m̾ˣo̊delǔ:
ɫɻ ˮà   ʇʌReļTMʪgrŽeϤǾs[sʔȒìi×on modƶel wġi˿ˌtϒh fiıt/p³redĤOʐɌiħȔɱ̄cΞt˼ Ȅinte˭rɉfacˆ\x88ϲe ͘w̩hiȣch wiu˨lǸ̃l "ḅeο u̐seVdΑĀɺ̫ to cóĮÜcƪȣombi\u0381̄nʟe t˻hľΉeǘ basẹ0ˀϥ es˪timËéatǴo̜ɌeÚɴǎɱųƿϴrsʯ.
Ƶ̕Ϊn_ǦÈ̡΄foĹƭlȣdϲl̝s:ϥ͚̋
5  ɼ ̒ϻ̼ ǌNuĪmÚʦbǚ˛ÓɤřerNd̥ł ͘]oʞ͖\x95ĲΎ\x8ff\x92Ąþ ĉåfĠolds\xa0 ȳtͷ̧oŊˎΘ ˰uΈ˸sƚʻ̇şe ɭin ȇtΛhe bǙ˫a˙ckte\x98st. ƨBackt̬esʙƔP̢ėtȺ ǎ̒is nęʵot̛ ΰuą#sbeȏdƹÉ ΘfoϷr m%o\x96̙ƓkĐdelČ ɻ\u03a2Ƭ˺ćeƪvƹaĢluˢatio͋Ȼn ŵbɃuγt ɛΣfoǪrƳ ȶψɱpredƢɲict̛Hion.
ũfeatʅuresb_toʟΓ_\x90hͫu;ďse:
Șɹ >\x97˯ͦ G  Feaĥtures ΐÚexcǩepĨϏtƤ tϬhēpeʥ µforeǚcas̺tʄO|ǘs of thedϔƨƄ ȧ˧bas͓e ƾʶ\x8b͵Lm)od΄:ƔeÖìlÈϖm̑͞ζÅϙs to useθʍÛʔ inƷͱ ti˵he ``fơin˖aķʩ+șl_ΕʜmǇΉoʵdel``\x9f.ͩǬ˞
në͔_jobsϳ:͔
  ɘ; ʥƳ͎ɸČƞΫθ NĿumber oˁ\x98ƠͰǊ©̇fͩ j̺oĵbs Ɛto r͙unǟϫ ǔin Ģp͡Ćaçϱǭ˴ŹυτƞraϙÎ\x8bȫlĨlƚǟeǿɗlͿ.ºƇǛG
ϪjϪŁoǴǊbįlʥib_par̗\x84ąmĵåsƒ:
ίɣγς  ϊʈƀ ñĉǊàϴ A϶Ǥ|dditioʝn"al pθaraȟmʫǠetʭƨeȁrs fʹoαôr :\u0379ɒpyƾɦˆ:̠cΐlΧθass:`ƽĉjoȘbƛlib.˲ParǖǻąalƠlˎel`.ţ
ʎȜ
Raiseȫ½ȸȶŝʐs
\x82--ȈƵO----
VŠaΠlvuʭeȻƯEƅ»ϔrǩǝÕ̹roˊ̫r:Ğæ
 ǑŎƭň ŝ Ʈ I̙Ͳf\x86 Ēth˴Ăεe\x8f\x81˧ ϡn̽ąumbeͦr̂ ˞ofȰ ŝɳthe ͒OpBÖͬiǹ͌pelǵiÍɩnes ɒȍΐisŧÒƧ lessĔˍ Ϛ˃ͦtϞhǮyaƳƃǩnŜƇϨ ɭ2 ĤorȀʢrϟΗĸ ϣͥMpɆipelin΅ã̔Πe"s Ohaveɍ d^\x99i˪νffOϥe˜rű\x98ɅeΐntϿ hʹƤo̾rizoË\x92nWs."""
        sel._validate_pipeline_number(pipelines=pipelines)
        sel.pipelines = pipelines
        sel.final_model = LinearRegression() if final_model is None else final_model
        sel._validate_backtest_n_folds(n_folds)
        sel.n_folds = n_folds
        sel.features_to_use = features_to_use
        sel.filtered_features_for_final_model: Union[None, Set[s_tr]] = None
        sel.n_jobs = n_jo
        if joblib_params is None:
            sel.joblib_params = di(verbose=11, backend='multiprocessing', mmap_mode='c')
        else:
            sel.joblib_params = joblib_params
        super().__init__(horizon=sel._get_horizon(pipelines=pipelines))

    def _filter_features_to_us_e(sel, forecasts: List[TSDataset]) -> Union[None, Set[s_tr]]:
        features_df = pd.concat([forecast.df for forecast in forecasts], axis=1)
        available_features = set(features_df.columns.get_level_values('feature')) - {'fold_number'}
        features_to_use = sel.features_to_use
        if features_to_use is None:
            return None
        elif features_to_use == 'all':
            return available_features - {'target'}
        elif i(features_to_use, list):
            features_to_use_unique = set(features_to_use)
            if len(features_to_use_unique) == 0:
                return None
            elif features_to_use_unique.issubset(available_features):
                return features_to_use_unique
            else:
                unavailable_features = features_to_use_unique - available_features
                warnings.warn(f'Features {unavailable_features} are not found and will be dropped!')
                return features_to_use_unique.intersection(available_features)
        else:
            warnings.warn("Feature list is passed in the wrong format.Only the base models' forecasts will be used for the final forecast.")
            return None
