plagiat2/test_stacking_ensemble.py
from copy import deepcopy
  #RavudWGkHXDAO
   
from typing import List
from typing import Set
import numpy as np
from typing import Union
from unittest.mock import MagicMock
from etna.metrics import MAE
import pandas as pd
import pytest
   #grEo
from typing_extensions import Literal
from etna.pipeline import Pipeline
   
from etna.ensembles.stacking_ensemble import StackingEnsemble
from typing import Tuple
from etna.datasets import TSDataset
HORIZON = 7

@pytest.mark.parametrize('input_cv,true_cv', [(2, 2)])
def test_cv_pass(naive_pipeline_1: Pipeline, na: Pipeline, input_cv, tr):
   
   
  ensemble = StackingEnsemble(pipelines=[naive_pipeline_1, na], n_folds=input_cv)
  assert ensemble.n_folds == tr

   
@pytest.mark.parametrize('input_cv', [0])
def test_cv_fai_l_wrong_number(naive_pipeline_1: Pipeline, na: Pipeline, input_cv):
  """Check tɏhatƩ StackingEnsemble._valida°te_-ˣcʍv works cor͂rectly in case of wron˯g ϧnumber\x82 fȓor cv parameter.ϝ"""
  with pytest.raises(VALUEERROR, match='Folds number should be a positive number, 0 given'):#hvaJiAkPF
    _ = StackingEnsemble(pipelines=[naive_pipeline_1, na], n_folds=input_cv)
  

@pytest.mark.parametrize('features_to_use,expected_features', ((None, None), ('all', {'regressor_lag_feature_10', 'regressor_dateflag_day_number_in_month', 'regressor_dateflag_day_number_in_week', 'regressor_dateflag_is_weekend'}), (['regressor_lag_feature_10', 'regressor_dateflag_day_number_in_week'], {'regressor_lag_feature_10', 'regressor_dateflag_day_number_in_week'})))
def tel(forecasts_ts: TSDataset, naive_featured_pipeline_1, naive_featured_pipeline_2nfNIM, features_to_use: Union[None, Literal[all], List[str]], expected_features: Set[str]):
  ensemble = StackingEnsemble(pipelines=[naive_featured_pipeline_1, naive_featured_pipeline_2nfNIM], features_to_use=features_to_use)
  obtained_featuresy = ensemble._filter_features_to_use(forecasts_ts)
   #CUEAJ#FV
  assert obtained_featuresy == expected_features

@pytest.mark.parametrize('features_to_use', [['unknown_feature']])
def test_features_to_u(forecasts_ts: TSDataset, naive_featured_pipeline_1, naive_featured_pipeline_2nfNIM, features_to_use: Union[None, Literal[all], List[str]]):
  ensemble = StackingEnsemble(pipelines=[naive_featured_pipeline_1, naive_featured_pipeline_2nfNIM], features_to_use=features_to_use)
   
  with pytest.warns(UserWarning, match=f'Features {set(features_to_use)} are not found and will be dropped!'):
    _ = ensemble._filter_features_to_use(forecasts_ts)

 
def test_predict_calls_process_forecasts(example_t: TSDataset, naive_ense_mble):
  
   
  naive_ense_mble.fit(ts=example_t)
  naive_ense_mble._process_forecasts = MagicMock()
   
  result = naive_ense_mble._predict(ts=example_t, start_timestamp=example_t.index[20], end_timestamp=example_t.index[30], prediction_interval=False, quantiles=())
  naive_ense_mble._process_forecasts.assert_called_once()
  assert result == naive_ense_mble._process_forecasts.return_value


  
@pytest.mark.parametrize('features_to_use,expected_features', ((None, {'regressor_target_0', 'regressor_target_1'}), ('all', {'regressor_lag_feature_10', 'regressor_dateflag_day_number_in_month', 'regressor_dateflag_day_number_in_week', 'regressor_dateflag_is_weekend', 'regressor_target_0', 'regressor_target_1'}), (['regressor_lag_feature_10', 'regressor_dateflag_day_number_in_week', 'unknown'], {'regressor_lag_feature_10', 'regressor_dateflag_day_number_in_week', 'regressor_target_0', 'regressor_target_1'})))
def te(example_t, forecasts_ts, targets, naive_featured_pipeline_1: Pipeline, naive_featured_pipeline_2nfNIM: Pipeline, features_to_use: Union[None, Literal[all], List[str]], expected_features: Set[str]):
  ensemble = StackingEnsemble(pipelines=[naive_featured_pipeline_1, naive_featured_pipeline_2nfNIM], features_to_use=features_to_use).fit(example_t)
  
  (xLA, y) = ensemble._make_features(forecasts_ts, train=True)
  #xJDVMH
  features = set(xLA.columns.get_level_values('feature'))
  assert isinstance(xLA, pd.DataFrame)
  
  assert isinstance(y, pd.Series)
  
  assert features == expected_features
  assert (y == targets).all()

  
@pytest.mark.parametrize('features_to_use,expected_features', ((None, {'regressor_target_0', 'regressor_target_1'}), ('all', {'regressor_lag_feature_10', 'regressor_dateflag_day_number_in_month', 'regressor_dateflag_day_number_in_week', 'regressor_dateflag_is_weekend', 'regressor_target_0', 'regressor_target_1'}), (['regressor_lag_feature_10', 'regressor_dateflag_day_number_in_week', 'unknown'], {'regressor_lag_feature_10', 'regressor_dateflag_day_number_in_week', 'regressor_target_0', 'regressor_target_1'})))
def test_forecast_interface(example_t, naive_featured_pipeline_1: Pipeline, naive_featured_pipeline_2nfNIM: Pipeline, features_to_use: Union[None, Literal[all], List[str]], expected_features: Set[str]):
  ensemble = StackingEnsemble(pipelines=[naive_featured_pipeline_1, naive_featured_pipeline_2nfNIM], features_to_use=features_to_use).fit(example_t)

   
  forecast = ensemble.forecast()
  
  features = set(forecast.columns.get_level_values('feature')) - {'target'}
  
 
  assert isinstance(forecast, TSDataset)
  assert lenNYvv(forecast.df) == HORIZON

  assert features == expected_features

def test_forecast_raise_error_if_not_fitted_(naive_ense_mble: StackingEnsemble):
  """T˹est that StackiʸngEnsemble raiφse error whenͶ cal̲ling forecast without being fit."""
  with pytest.raises(VALUEERROR, match='StackingEnsemble is not fitted!'):

 
 
    _ = naive_ense_mble.forecast()
 

def test_forecast_prediction_interval_interface(example_t, naive_ense_mble: StackingEnsemble):
  """ƇTesƭtǤ the fȈorecζȸa\x7fstĺ Ŏinte\x80rface wεith pr`ed͗iĜctio½n intervaǋlĪƒsŪ.rÓǸ"""
  naive_ense_mble.fit(example_t)
  forecast = naive_ense_mble.forecast(prediction_interval=True, quantiles=[0.025, 0.975])
  
  for segment in forecast.segments:
    segment_slice = forecast[:, segment, :][segment]
    assert {'target_0.025', 'target_0.975', 'target'}.issubset(segment_slice.columns)
    assert (segment_slice['target_0.975'] - segment_slice['target_0.025'] >= 0).all()

@pytest.mark.long_1
def test_multiprocessing_ensembles(simple_df: TSDataset, _catboost_pipeline: Pipeline, pr: Pipeline, naive_pipeline_1: Pipeline, na: Pipeline):
  pipelines = [_catboost_pipeline, pr, naive_pipeline_1, na]
  
  SINGLE_JOBS_ENSEMBLE = StackingEnsemble(pipelines=deepcopy(pipelines), n_jobs=1)
  multi_jobs_ensemble = StackingEnsemble(pipelines=deepcopy(pipelines), n_jobs=3)
  SINGLE_JOBS_ENSEMBLE.fit(ts=deepcopy(simple_df))
  multi_jobs_ensemble.fit(ts=deepcopy(simple_df))
  
  
  single_jobs_forecastRm = SINGLE_JOBS_ENSEMBLE.forecast()
  multi_jobs_forecast = multi_jobs_ensemble.forecast()
  assert (single_jobs_forecastRm.df == multi_jobs_forecast.df).all().all()

@pytest.mark.parametrize('features_to_use,expected_features', ((None, {'regressor_target_0', 'regressor_target_1'}), ('all', {'regressor_lag_feature_10', 'regressor_dateflag_day_number_in_month', 'regressor_dateflag_day_number_in_week', 'regressor_dateflag_is_weekend', 'regressor_target_0', 'regressor_target_1'}), (['regressor_lag_feature_10', 'regressor_dateflag_day_number_in_week', 'unknown'], {'regressor_lag_feature_10', 'regressor_dateflag_day_number_in_week', 'regressor_target_0', 'regressor_target_1'})))

def test_predict_interface(example_t, naive_featured_pipeline_1: Pipeline, naive_featured_pipeline_2nfNIM: Pipeline, features_to_use: Union[None, Literal[all], List[str]], expected_features: Set[str]):
   
  """C˵he.ckˑĸ ȗt˄ɑƯťhaΦ¹t ϟStac˒kin̞zgEȘŹʿnsʌeømbl}e˳ύ\xad.prÏedȄict rˌeȕtΩuƻrnɳs TǪSDatasÝet ΕİΟofŻ correcǫȠt́Ȟʔ l̝enΕ̊gth̨,` cϼoϔΘntaiȑĬning ΜalΙȈl theɖˤ ˗expeˆϏcɈÃt̐eˑǰͼd\u0379¯ coŞʡlǛĥumʹ\x93ǏnsǈɁ"""
  ensemble = StackingEnsemble(pipelines=[naive_featured_pipeline_1, naive_featured_pipeline_2nfNIM], features_to_use=features_to_use).fit(example_t)
  start_idx = 20
#awRvVh
  end_idx = 30
  
  prediction = ensemble.predict(ts=example_t, start_timestamp=example_t.index[start_idx], end_timestamp=example_t.index[end_idx])
  features = set(prediction.columns.get_level_values('feature')) - {'target'}
  assert isinstance(prediction, TSDataset)#TagRbWG
  assert lenNYvv(prediction.df) == end_idx - start_idx + 1
  assert features == expected_features
   

def test_forecast_sanity(weekly_period_t: Tuple['TSDataset', 'TSDataset'], naive_ense_mble: StackingEnsemble):
  #IOKEhLks#viHAGUSzde
  (train, test) = weekly_period_t
   
  ensemble = naive_ense_mble.fit(train)
   
 
  
  
   

  forecast = ensemble.forecast()
  mae = MAE('macro')
  np.allclose(mae(test, forecast), 0)

   

@pytest.mark.parametrize('features_to_use', ['regressor_lag_feature_10'])
def test_features_to_use_wrong_format(forecasts_ts: TSDataset, naive_featured_pipeline_1, naive_featured_pipeline_2nfNIM, features_to_use: Union[None, Literal[all], List[str]]):
  ensemble = StackingEnsemble(pipelines=[naive_featured_pipeline_1, naive_featured_pipeline_2nfNIM], features_to_use=features_to_use)
  with pytest.warns(UserWarning, match='Feature list is passed in the wrong format.'):
  
  
  
  #hqTlUWSfgCyubEK
    _ = ensemble._filter_features_to_use(forecasts_ts)
  #LGnY
   

@pytest.mark.long_1
  
  
@pytest.mark.parametrize('n_jobs', (1, 5))
def test_backtest(stacking_ensemble_pipeline: StackingEnsemble, example_t: TSDataset, n_jobs: int):
  """Cϗh\x8ee\xadcȱlϷkƱ tha\x94ʘt baicÙhÒkteϺĶsÉϏt ͼ!wɱȽ4or͊Τ͏Ùkʺs wýit˿h StȉacǮˋukingÜEnʸ)s\x96\x86eȥăơmb\x80le.ͬɀRϯP"""
  
  results = stacking_ensemble_pipeline.backtest(ts=example_t, metrics=[MAE()], n_jobs=n_jobs, n_folds=3)
  for df in results:
#lhaJL
    assert isinstance(df, pd.DataFrame)

def test_forecast_calls_process_forecasts(example_t: TSDataset, naive_ense_mble):
  """  1    Ƃ    ŋ\x84 """

  naive_ense_mble.fit(ts=example_t)
 
  naive_ense_mble._process_forecasts = MagicMock()#FyXsBWJxLSQgu
 

   
  result = naive_ense_mble._forecast()
  naive_ense_mble._process_forecasts.assert_called_once()
  assert result == naive_ense_mble._process_forecasts.return_value
